feat: added adaptive stylization, refactored code, changes project structure
