+++ public/style.scss
@@ -217,31 +217,25 @@ code {
 .book__details__card {
   display: flex;
   flex-direction: row;
-  padding: 20px;
-  border: '1px solid #ddd';
-  border-radius: 5px;
-
-  width: '400px';
-  margin: '20px auto';
-  justify-content: space-around;
   .book__details__thumbnail {
-    width: 600px;
-    height: 400px;
+    width: 33%;
+    height: 430px;
     margin-right: 20px;
-    background-color: #f9f9f9;
+    background-color: rgb(240, 240, 240);
+    text-align: center;
   }
 }
-.img__container {
-  width: 300px;
-  height: 200px;
-  margin-top: 50px;
-  margin-left: 100px;
+.book__details__thumbnail {
   img {
+    width: 40%;
     object-fit: cover;
-    width: 400px;
-    height: 100%;
+    transform: translateY(8%);
+    vertical-align: center;
+    border-radius: 5px;
+    box-shadow: 10px 10px 20px rgba(0, 0, 0, 0.2);
   }
 }
+.book__details__card {
 .book__info {
   display: flex;
   flex-direction: column;
@@ -249,6 +243,7 @@ code {
   background-color: #fff;
   justify-content: flex-start;
   align-items: flex-start;
+  margin-left: 50px;
   .category {
     font-size: 14px;
     color: gray;
@@ -258,6 +253,7 @@ code {
     font-size: 28px;
     font-weight: bold;
     margin: 10px 0;
+    margin-top: 30px;
   }
   .author {
     font-size: 16px;
@@ -274,4 +270,317 @@ code {
     width: 600px;
     height: 150px;
   }
+}
+}
+.book__card {
+  height: 100%;
+  border-radius: 5px;
+}
+ 
+@media (min-width: 768px) {
+  .book__details__card {
+      flex-direction: row;
+      align-items: flex-start;
+  }
+  .book__info {
+      text-align: left;
+      margin-left: 20px;
+  }
+}
+.buttonBack {
+  margin-left: -600px;
+  margin-top: 380px;
+  padding: 10px 20px;
+  font-size: 16px;
+  color: #fff;
+  background-color: #007bff;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+}
+.load__more {
+  padding: 10px 20px;
+  font-size: 16px;
+  color: #fff;
+  background-color: #007bff;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+
+}
+.buttonBack:hover {
+  background-color: #0056b3;
+}
+/* Extra small devices (phones, 320px and up) */
+@media (min-width: 320px)  and (max-width:479px) {
+  .book__info .title {
+    font-size: 22px;
+  }
+  .book__info .description {
+    max-width: 250px;
+  }
+  .book__details__card {
+    justify-content: center;
+    align-items: center;
+    margin-top: 15px;
+    border-radius: 5px;
+    .category {
+      width: 45%;
+      margin: 0;
+      margin-left: 55px;
+      margin-top: 20px;
+    }
+    .buttonBack {
+      margin-top: 50px;
+      margin-left: 0px;
+      margin-right: 45px;
+    }
+    .book__details__thumbnail {
+      width: 100%;
+      height: 170px;
+      margin: 0;
+      img {
+        height: 75%;
+        width: 30%;
+        transform: translateY(10%);
+      }
+    }
+  }
+  .searchButton {
+    right: 75px;
+  }
+  .search__wrap {
+    margin-top: 75px;
+  }
+  .book__card {
+    max-width: calc(75% - 10px);
+    flex: 1 1 calc(75% - 16px);
+  }
+}
+/* Small devices (phones, 480px and up) */
+@media (min-width: 480px) and (max-width: 767px) {
+  .book__info .title {
+    font-size: 24px;
+  }
+  .book__info .description {
+    max-width: 400px;
+  }
+  .book__details__card {
+    margin-top: 15px;
+    border-radius: 5px;
+    .buttonBack {
+      margin-top: 50px;
+      margin-left: -250px;
+    }
+    .book__details__thumbnail {
+      height: 250px;
+      width: 50%;
+    }
+  }
+  .searchButton {
+    right: 125px;
+  }
+  .book__card {
+    max-width: calc(75% - 30px);
+    flex: 1 1 calc(75% - 36px);
+  }
+}
+/* Medium devices (tablets, 768px and up) */
+@media (min-width: 768px) and (max-width: 1023px) {
+  .book__details__card {
+    flex-direction: row;
+    align-items: flex-start;
+    text-align: left;
+  }
+  .book__card {
+    max-width: calc(50% - 50px);
+    flex: 1 1 calc(50% - 56px);
+  }
+  .book__details__thumbnail {
+    width: 33%;
+    margin-right: 20px;
+    height: 430px;
+  }
+  .book__info {
+    align-items: flex-start;
+  }
+  .book__info .description {
+    margin-left: 0;
+    max-width: 500px;
+  }
+  .searchButton {
+    right: 245px;
+  }
+  .book__details__card {
+    .buttonBack {
+      margin-right: -75px;
+      margin-top: 80px;
+    }
+    .category {
+      width: 50%;
+    }
+  .book__details__thumbnail {
+     width: 50%;
+     border-radius: 5px;
+     margin-top: 15px;
+     img {
+      width: 40%;
+      height: 56%;
+      transform: translateY(38%);
+     }
+  }
+}
+}
+/* Large devices (desktops, 1024px and up) */
+@media (min-width: 1024px) and (max-width:1199px) {
+  .book__info .title {
+    font-size: 28px;
+  }
+  .book__info .description {
+    max-width: 500px;
+    
+  }
+  .book__details__card {
+    .category {
+       max-width: 50%;
+    }
+  }
+  .searchButton {
+    right: 300px;
+  }
+  .book__card {
+    max-width: calc(33% - 70px);
+    flex: 1 1 calc(33% - 76px);
+  }
+}
+/* Extra large devices (large desktops, 1200px and up) */
+@media (min-width: 1200px) and (max-width: 1400px) {
+  #root {
+    overflow: hidden;
+  }
+  .container {
+    padding: 40px;
+  }
+  .book__details__thumbnail {
+    height: 450px;
+  }
+  .searchButton {
+    right: 386px;
+  }
+  .book__details__card {
+    .category {
+       max-width: 50%;
+    }
+  }
+  .book__card {
+    max-width: calc(33% - 70px);
+    flex: 1 1 calc(33% - 76px);
+  }
+}
+@media (max-width: 1200px) {
+  .header__wrap {
+    padding: 50px 0;
+  }
+
+  .searchbar {
+    max-width: 100%;
+  }
+}
+
+@media (max-width: 992px) {
+  .book__details__card {
+    flex-direction: column;
+    align-items: center;
+  }
+  .book__info {
+    align-items: center;
+  }
+
+  .description {
+    width: 100%;
+    max-width: none;
+  }
+
+  .Filter {
+    flex-direction: column;
+  }
+
+  .sorting__wrap {
+    margin-left: 0;
+    margin-top: 10px;
+  }
+}
+
+@media (max-width: 768px) {
+  .header__wrap {
+    padding: 30px 0;
+  }
+
+  .searchbar {
+    width: 100%;
+    max-width: 300px;
+  }
+
+  .input__container {
+    width: 100%;
+  }
+
+  .Filter {
+    flex-direction: column;
+    align-items: center;
+  }
+
+  .category__wrap, .sorting__wrap {
+    margin-top: 10px;
+    justify-content: center;
+  }
+}
+
+@media (max-width: 576px) {
+  .header__wrap {
+    padding: 20px 0;
+  }
+
+  .h1__container {
+    margin-bottom: 10px;
+  }
+
+  .searchbar {
+    max-width: 250px;
+  }
+
+  .Filter {
+    width: 100%;
+    justify-content: center;
+  }
+}
+
+@media (max-width: 480px) {
+  .book__info .title {
+    font-size: 24px;
+  }
+
+  .book__info .author {
+    font-size: 14px;
+  }
+
+  .book__info .description {
+    font-size: 12px;
+  }
+}
+
+@media (max-width: 320px) {
+  .header__wrap {
+    padding: 10px 0;
+  }
+
+  .h1__container h1 {
+    font-size: 24px;
+  }
+
+  .searchbar {
+    width: 100%;
+    max-width: 200px;
+  }
+
 }
